Refactor tunnel state machine
