+++ nym-vpn-core/crates/nym-vpn-lib/src/platform/mod.rs
@@ -27,9 +27,9 @@ use nym_vpn_store::mnemonic::MnemonicStorage as _;
 
 use self::error::VpnError;
 #[cfg(target_os = "android")]
-use crate::tunnel_state_machine::tunnel::wireguard::android::AndroidTunProvider;
+use crate::tunnel_state_machine::tunnel_provider::android::AndroidTunProvider;
 #[cfg(target_os = "ios")]
-use crate::tunnel_state_machine::tunnel::wireguard::ios::tun_provider::OSTunProvider;
+use crate::tunnel_state_machine::tunnel_provider::ios::OSTunProvider;
 use crate::{
     gateway_directory::GatewayClient,
     tunnel_state_machine::{
