--- nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/wireguard/ios/mod.rs
@@ -1,4 +1,0 @@
-pub mod default_path_observer;
-pub mod dns64;
-pub mod tun;
-pub mod tun_provider;
