+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel_provider/ios/default_path_observer.rs
@@ -1,3 +1,6 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
 use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
 
 /// The type alias for the receiving end of path observer.
